It looks like the git container is working

+++ docker-git/docker-entrypoint.sh
@@ -11,7 +11,7 @@ GIT_PASSWD_FILENAME=git.passwd
 # ************************************************************
 # git               : Starts apache running. This is the containers default
 # git_backup        : archives the git repositories into the IMPORT_EXPORT_PATH
-# git_restore       : restore all repositories inside the IMPORT_EXPORT_PATH
+# git_restore       : restore repository archive names (*.backup.tar.gz | *.mirror.tar.gz) from the remining arguments
 # ssl_generate      : generates a self signed certificate authority
 # passwd_generate   : creates an initial httpasswd file with the novatech user
 # new_repository    : creates a new bare repository named from remaining arguments
@@ -20,40 +20,50 @@ case ${1} in
     'git')
         # Apache gets grumpy about PID files pre-existing
         rm -f /var/run/apache2/apache2.pid
+        # if apache dose not shut down properly and release all locks
+        rm -f /var/lock/apache2/DAVLock
         # Start apache
         exec apache2 -D FOREGROUND
         ;;
 
     git_backup)
+        workdir=$(mktemp -d /tmp/git_backup.XXXXXXXXXX)
         # commands export the GIT repositories for backup
         for repo_path in ${GIT_BASE_DIR}/*
         do
             repo_name=$(basename ${repo_path})
-            if [[ ! -f ${repo_path}/config ]] ; then
-                continue
-            fi
-            if [[ -f ${IMPORT_EXPORT_PATH}/${repo_name} ]] ; then
-                rm -f ${IMPORT_EXPORT_PATH}/${repo_name}
-            fi
-            git clone --mirror ${repo_path} ${IMPORT_EXPORT_PATH}/${repo_name}
+            [[ ! -f ${repo_path}/config ]] && continue
+            # make a backup of the working repository
+            pushd ${repo_path}
+            tar -czf ${IMPORT_EXPORT_PATH}/${repo_name}.backup.tar.gz *
+            popd
+            # (If there is a commit at the same time this is happening this can leave the repository in a bad state)
+            # Therefore: mirror repository to working directory and archive the mirror repository
+            git clone --mirror ${repo_path} ${workdir}/${repo_name}
+            pushd ${workdir}/${repo_name}
+            tar -czf ${IMPORT_EXPORT_PATH}/${repo_name}.mirror.tar.gz *
+            popd
         done
         ;;
 
     git_restore)
-        # Remove any git repositories currently in existance
-        [[ ! -d ${GIT_BASE_DIR} ]] && rm -rf ${GIT_BASE_DIR}/*.git
-        # Import mirrored GIT repositories
-        for backup_repo_path in ${IMPORT_EXPORT_PATH}/*.git
+        shift
+        ARCHIVED_REPOSITORIES=(${*})
+        for archive_file in ${ARCHIVED_REPOSITORIES[*]}
         do
-            if [[ ! -d ${backup_repo_path} ]] ; then
+            if [[ ! -e ${IMPORT_EXPORT_PATH}/${archive_file} ]] ; then
+                echo "SKIPPING (Could not locate): ${IMPORT_EXPORT_PATH}/${archive_file}"
                 continue
             fi
-            repo_name=$(basename ${backup_repo_path})
-            if [[ -d ${GIT_BASE_DIR}/${repo_name} ]] ; then
-                rm -rf ${GIT_BASE_DIR}/${repo_name}
-            fi
-            git clone --mirror ${backup_repo_path} ${GIT_BASE_DIR}/${repo_name}
+            repo_name=$(echo ${archive_file} | sed 's/\.backup\.tar\.gz$//' | sed 's/\.mirror\.tar\.gz$//')
+            echo "[RESTORE REPOSITORY] : ${repo_name}"
+            [[ ! -e ${GIT_BASE_DIR}/${repo_name} ]] && rm -rf ${GIT_BASE_DIR}/${repo_name}
+            mkdir ${GIT_BASE_DIR}/${repo_name}
+            tar -xzf ${IMPORT_EXPORT_PATH}/${archive_file} --directory=${GIT_BASE_DIR}/${repo_name}
             pushd ${GIT_BASE_DIR}/${repo_name}
+            touch git-daemon-export-ok
+            cp hooks/post-update.sample hooks/post-update
+            git config http.receivepack true
             git update-server-info
             popd
         done
@@ -112,6 +122,9 @@ case ${1} in
             mkdir ${GIT_BASE_DIR}/${repo_name}.git
             pushd ${GIT_BASE_DIR}/${repo_name}.git
             git init --bare --shared=group
+            touch git-daemon-export-ok
+            cp hooks/post-update.sample hooks/post-update
+            git config http.receivepack true
             git update-server-info
             popd
         done
